wip: validation edit/create form collection
