+++ src/pages/collection/create.vue
@@ -1,16 +1,24 @@
 <script setup>
 import { useRouter } from "vue-router";
 import { onMounted } from "vue";
+import { Form } from "vee-validate";
+import * as yup from "yup";
 import FormCollection from "@/components/FormCollection.vue";
 import useFormCollection from "@/composables/useFormCollection";
 
 const router = useRouter();
 const { addNewCollection, form, resetForm } = useFormCollection();
+const schema = yup.object({
+  name: yup.string().required(),
+  description: yup.string(),
+});
 
-function handleSubmit() {
+const onSubmit = (values) => {
+  form.value = values;
+  form.value.todos = [];
   addNewCollection();
   router.push("/");
-}
+};
 
 onMounted(() => {
   resetForm();
@@ -18,6 +26,12 @@ onMounted(() => {
 </script>
 <template>
   <div class="mx-auto max-w-2xl">
-    <FormCollection :form="form" @handleSubmit="handleSubmit" />
+    <Form
+      @submit="onSubmit"
+      :validation-schema="schema"
+      class="flex flex-col gap-4"
+    >
+      <FormCollection :form="form" />
+    </Form>
   </div>
 </template>
